Add types, update __repr__  and improve docs for data_store and models
Add types for data_store and model, as part of an effort to gradually
roll out types in the code base. Related to #56.

Change import structure for sqlalchemy and only import the top level
package, to make the code easier to read and understand where e.g.,
"Column" comes from.

Improve the docstrings by attempting to have a first line summary and
additional description in follow lines. Types were removed from the
docstring as they are part of the function signature.
The docstring is written in 120 character lines, given that the codebase
is using them, but it would be easier to read in terminal usage if they
were 80 char lines.

Update the __repr__ method of models to print only the relevant files
(and remove a bunch of unnecessary code).